// model/border.ts
// https://www.compart.com/en/unicode/block/U+2500
// ┏━┳┓
// ┃━━┃
// ┣━╋┫
// ┗━┻┛

export const enum Border {
  Horizontal = "━",
  HorizontalDown = "┳",
  HorizontalUp = "┻",
  Vertical = "┃",
  VerticalRight = "┣",
  VerticalLeft = "┫",
  TopRight = "┏",
  TopLeft = "┓",
  BottomRight = "┗",
  BottomLeft = "┛",
  Cross = "╋",
}